application/backend/requirements.txt
flask
flask-cors
psycopg2-binary
redis
boto3
gunicorn
python-dotenv
